Extract meal planner derived view models
